+++ smartContract/oracle.go
@@ -72,19 +72,6 @@ type lockAndUnlockResult struct {
 	Result  bool   `json:"result"`
 }
 
-type listWalletsResult struct {
-	Jsonrpc string `json:"jsonrpc"`
-	Id      int    `json:"id"`
-	Result  []struct {
-		Url      string `json:"url"`
-		Status   string `json:"status"`
-		Accounts []struct {
-			Address string `json:"address"`
-			Url     string `json:"url"`
-		} `json:"accounts"`
-	} `json:"result"`
-}
-
 // 定义仅本包内可见的数据
 var (
 	// oracle对象
@@ -135,14 +122,8 @@ func NewOracle() OracleWriter {
 		taskMap = new(oracleTaskMap)
 		taskMap.jobMap = make(map[string]string)
 
-		// 在这里先尝试解锁一次账户
-		err := oracle.tryUnlockAccount()
-		if err != nil {
-			logger.Fatal("尝试解锁账户失败")
-		}
-
 		// 开始监听请求智能合约
-		err = oracle.registerOracleRequestContractMonitor()
+		err := oracle.registerOracleRequestContractMonitor()
 		if err != nil {
 			logger.Fatal("监听OracleRequestContract失败")
 		} else {
@@ -162,6 +143,13 @@ func (o *oracleClient) WriteData(jobID string, data string) (bool, error) {
 		logger.Println("尝试解锁账户失败")
 		return false, err
 	}
+	// 写入数据成功后，尝试重新锁定账户
+	defer func(o *oracleClient) {
+		err := o.tryLockAccount()
+		if err != nil {
+			logger.Println("尝试重新锁定账户失败")
+		}
+	}(o)
 
 	// 获取私钥，该私钥是oracle的私钥
 	privateKey, err := crypto.HexToECDSA(o.OraclePrivateKey)
@@ -398,35 +386,6 @@ func (o *oracleClient) tryUnlockAccount() error {
 		return err
 	}
 
-	listWalletRequest := "{\"jsonrpc\":\"2.0\",\"method\":\"personal_listWallets\",\"params\":[],\"id\":1}"
-	listWalletResultBytes, err := invokeJsonRpc(o.EthHttpUrl, []byte(listWalletRequest))
-	if err != nil {
-		return err
-	}
-
-	listWalletResult := new(listWalletsResult)
-	err = json.Unmarshal(listWalletResultBytes, listWalletResult)
-	if err != nil {
-		return err
-	}
-
-	isLocked := true
-	for _, result := range listWalletResult.Result {
-		for _, account := range result.Accounts {
-			if account.Address == publicKeyAddress.Hex() {
-				if result.Status == "Locked" {
-					isLocked = true
-					break
-				}
-			}
-		}
-	}
-
-	if !isLocked {
-		// 没有被锁住，那么就不进行任何操作
-		return nil
-	}
-
 	// 如果被锁住了，那么执行如下解锁账户的操作
 	unlockRequest := "{\"jsonrpc\":\"2.0\",\"method\":\"personal_unlockAccount\",\"params\":[\"%s\", \"%s\", 30],\"id\":1}"
 	param := fmt.Sprintf(unlockRequest, publicKeyAddress.Hex(), o.OracleAccountPasswd)
@@ -443,6 +402,36 @@ func (o *oracleClient) tryUnlockAccount() error {
 	}
 
 	return fmt.Errorf("解锁账户失败")
+}
+
+// 尝试锁定账户
+func (o *oracleClient) tryLockAccount() error {
+	privateKey, err := crypto.HexToECDSA(o.OraclePrivateKey)
+	if err != nil {
+		return err
+	}
+	err, publicKeyAddress := getPublicKeyAddress(privateKey)
+	if err != nil {
+		return err
+	}
+	lockRequest := "{\"jsonrpc\":\"2.0\",\"method\":\"personal_lockAccount\",\"params\":[\"%s\"],\"id\":1}"
+	param := fmt.Sprintf(lockRequest, publicKeyAddress.Hex())
+	lockResultBytes, err := invokeJsonRpc(o.EthHttpUrl, []byte(param))
+	if err != nil {
+		return err
+	}
+
+	result := new(lockAndUnlockResult)
+	err = json.Unmarshal(lockResultBytes, result)
+	if err != nil {
+		return err
+	}
+
+	if result.Result {
+		return nil
+	}
+
+	return fmt.Errorf("锁定账户失败")
 }
 
 // 生成ethRequest，Marshal 成[]byte ，传入do函数即可操作ethereum 节点
